requirements.txt
streamlit==1.32.0
torch==2.1.1
torchvision==0.16.1
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
